refactor : footer Recipe 분리
