+++ App/src/js/recipes/footer/FooterInfoItemRecipe.js
@@ -1,0 +1,17 @@
+export function FooterInfoItemRecipe({ href, textContent }) {
+  return {
+    tagName: "li",
+    attrs: {
+      class: "footer__info__item",
+    },
+    children: [
+      {
+        tagName: "a",
+        attrs: {
+          href,
+        },
+        textContent,
+      },
+    ],
+  };
+}
