+++ server/db.json
@@ -552,6 +552,10 @@
     {
       "keyword": "다다다",
       "id": 6
+    },
+    {
+      "keyword": "가나다라마",
+      "id": 7
     }
   ]
 }
